Implement polkit authorization for Apply method
When the `com.ubuntu.ProxyManager.Apply` method is called, before
changing any proxy settings, check if the caller is authorized to
perform the action.

We first get the credentials of the user based on the D-Bus sender
information - this is accomplished by updating the signature of our
Apply method to add a `dbus.Sender` field. This doesn't affect our
exported method signature, as godbus will treat this as a special type,
setting its value to the sender of the message.

With our sender information in place, we can query for the UID and PID
behind the action, using the `org.freedesktop.DBus` interface to call
the `GetConnectionCredentials` method.

After we get the UID and PID, we forward them to a polkit call that
determines whether the caller is authorized to perform the action.

If the caller is not authorized, we return early and close the program
with an error.

DEENG-591