Allow configuration of environment system proxy
The actual application of the environment proxy settings is quite
simple, the more interesting part is the abstraction of the proxy
settings so we can easily apply them to other backends in the
future (like APT and GSettings).

We validate all proxy settings to ensure they're in the right format, we
properly escape them while ensuring we don't escape the already escaped
ones, then we can represent them in the required format which in this
case is as environment variables.

DEENG-553