Allow configuration of GSettings proxy
Similar to APT and environment proxy, GSettings configuration is done by
writing to a file, with the addition of executing `glib-compile-schemas`
afterwards in order to make GSettings aware of the changes.

Additionally, there's more parsing involved here due to the specific
format required the override files, and the `org.gnome.system.proxy`
GSchema.

Because of the need to execute a shell command, more intricacies are
involved, such as executing the command if we are asked to remove all
settings, and handling command execution failure by reverting to the
previous version of the GSchema override file.

The ubuntu-proxy-manager service is designed to run on Ubuntu machines,
but there may be valid cases where users would like to run it on other
flavors that aren't necessarily GNOME-based - so if we can't find the
`glib-compile-schemas` command, we just log a message and cleanly exit.

DEENG-561