Not all protocols are supported by all backends
When checking if we should remove the proxy configuration files, we
should only consider protocols that are supported by the backend.

For example, APT doesn't support no_proxy or autoconfiguration, yet if
one of the options were to be set in the proxy manager, we would still
write an empty APT configuration.