Handle multiple concurrent D-Bus calls of the app
This commit reworks the lifecycle management of the application in a
separate package, moving away from a context cancelling approach in
favor of handling multiple runs using waitgroups.

The app mutex is now the responsibility of the lifecycle package as
well, abstracting all of the functionality outside of the app package.

To manage the lifecycle of the app, we attach a Lifecycle instance to
the proxyManagerBus object.

We call Start() when a client calls our D-Bus function, which increments
the lifecycle counter and causes subsequent calls to block until the
current one finishes.

When a run finishes we call RunDone() which signals the lifecycle to
decrement the counter and start processing the next run if applicable.

Once the counter reaches zero, the Wait() function returns a joined
representation of all errors that occurred during the runs, thus
effectively ending the program.

DEENG-602