Handle multiple concurrent D-Bus calls of the app
This commit reworks the lifecycle management of the application, moving
away from a context cancelling approach in favor of handling multiple
runs using channels. D-Bus Apply calls are queued in the applyCalls
channel, and responses are received from the applyResponse channel.

We no longer need an app mutex, as we process apply calls in a queue so
there's no possibility of multiple calls running concurrently.

DEENG-602